+++ src/DuckGame/Network/3NetworkCore.cs
@@ -22,7 +22,7 @@ namespace DuckGame
         public NMConnect(byte received, NetIndex4 s, string v, string mH)
         {
             this.version = v;
-            this.connectsReceived = (NetIndex4)(int)received;
+            this.connectsReceived = (NetIndex4)received;
             this.remoteSession = s;
             this.modHash = mH;
         }
