+++ src/DuckGame/Network/GhostManagerState.cs
@@ -10,7 +10,7 @@ namespace DuckGame
     public class GhostManagerState : Thing
     {
         public StateBinding _predictionIndexBinding = new StateBinding(nameof(predictionIndex));
-        public NetIndex16 predictionIndex = new NetIndex16((int)short.MaxValue);
+        public NetIndex16 predictionIndex = new NetIndex16(short.MaxValue);
 
         public GhostManagerState()
           : base()
