+++ src/DuckGame/Stuff/CryoMonitor.cs
@@ -15,7 +15,7 @@ namespace DuckGame
           : base(xpos, ypos)
         {
             this.graphic = new Sprite("survival/cryoMonitor");
-            this.center = new Vec2((float)(this.graphic.w / 2), (float)(this.graphic.h / 2));
+            this.center = new Vec2(this.graphic.w / 2, this.graphic.h / 2);
             this._collisionSize = new Vec2(32f, 32f);
             this._collisionOffset = new Vec2(-16f, -16f);
             this.depth = (Depth)0.9f;
