+++ src/DuckGame/Stuff/Drumset/LowTom.cs
@@ -15,9 +15,9 @@ namespace DuckGame
           : base(xpos, ypos)
         {
             this.graphic = new Sprite("drumset/largeTom");
-            this.center = new Vec2((float)(this.graphic.w / 2), (float)(this.graphic.h / 2));
+            this.center = new Vec2(this.graphic.w / 2, this.graphic.h / 2);
             this._stand = new Sprite("drumset/lowTomStand");
-            this._stand.center = new Vec2((float)(this._stand.w / 2), 0.0f);
+            this._stand.center = new Vec2(this._stand.w / 2, 0.0f);
             this._sound = "lowTom";
         }
