+++ src/DuckGame/Tiles/TutorialSign.cs
@@ -15,7 +15,7 @@ namespace DuckGame
             if (image == null)
                 return;
             this.graphic = new Sprite(image);
-            this.center = new Vec2((float)(this.graphic.w / 2), (float)(this.graphic.h / 2));
+            this.center = new Vec2(this.graphic.w / 2, this.graphic.h / 2);
             this._collisionSize = new Vec2(16f, 16f);
             this._collisionOffset = new Vec2(-8f, -8f);
             this.depth = - 0.5f;
