+++ src/DuckGame/Tiles/TutorialSign05.cs
@@ -17,7 +17,7 @@ namespace DuckGame
 
         public override void Draw()
         {
-            Color color = new Color((int)sbyte.MaxValue, (int)sbyte.MaxValue, (int)sbyte.MaxValue);
+            Color color = new Color(sbyte.MaxValue, sbyte.MaxValue, sbyte.MaxValue);
             Graphics.DrawString("@JUMP@", new Vec2(this.x - 26f, this.y + 32f), Color.White * 0.5f);
             Graphics.DrawString("@JUMP@", new Vec2(this.x - 5f, this.y - 16f), Color.White * 0.5f);
             Graphics.DrawString("@JUMP@", new Vec2(this.x + 15f, this.y - 8f), Color.White * 0.5f);
