+++ src/MonoTime/Debug.cs
@@ -28,15 +28,15 @@ namespace DuckGame
         [Conditional("DEBUG")]
         public static void DrawLine(Vec2 p1, Vec2 p2, Color col, float width = 1f)
         {
-            p1 = new Vec2((float)(int)p1.x, (float)(int)p1.y);
-            p2 = new Vec2((float)(int)p2.x, (float)(int)p2.y);
-            float rotation = (float)Math.Atan2((double)p2.y - (double)p1.y, (double)p2.x - (double)p1.x);
+            p1 = new Vec2((int)p1.x, (int)p1.y);
+            p2 = new Vec2((int)p2.x, (int)p2.y);
+            float rotation = (float)Math.Atan2(p2.y - (double)p1.y, p2.x - (double)p1.x);
             float length = (p1 - p2).length;
             DuckGame.Graphics.Draw((Tex2D)Debug._blank, p1, new Rectangle?(), col, rotation, Vec2.Zero, new Vec2(length, width), SpriteEffects.None, (Depth)1f);
         }
 
         [Conditional("DEBUG")]
-        public static void DrawRect(Vec2 p1, Vec2 p2, Color col) => DuckGame.Graphics.Draw((Tex2D)Debug._blank, p1, new Rectangle?(), col, 0.0f, Vec2.Zero, new Vec2((float)-((double)p1.x - (double)p2.x), (float)-((double)p1.y - (double)p2.y)), SpriteEffects.None, (Depth)1f);
+        public static void DrawRect(Vec2 p1, Vec2 p2, Color col) => DuckGame.Graphics.Draw((Tex2D)Debug._blank, p1, new Rectangle?(), col, 0.0f, Vec2.Zero, new Vec2((float)-(p1.x - (double)p2.x), (float)-(p1.y - (double)p2.y)), SpriteEffects.None, (Depth)1f);
 
         [Conditional("DEBUG")]
         [Conditional("SWITCH")]
